Fix the lockfile not being respected due to not checking for INVALID_HANDLE_VALUE. Refactor the main updater code a bit, primarily so that the lockfile's ScopedHandle is naturally destroyed before starting the launcher (required, as the lockfile fix by itself caused issues otherwise).

+++ self_updater/self_updater.cpp
@@ -103,6 +103,57 @@ HANDLE Updater::TryOpenLauncherProcessHandle(int argc, char** argv) {
 	return launcherHandle;
 }
 
+bool Updater::TryDeleteOldRenamedUpdaterExe() {
+	int attempts = 0;
+	while (std::filesystem::exists(RenamedUpdaterExeFilename) && !DeleteFileA(RenamedUpdaterExeFilename)) {
+		attempts++;
+		if (attempts == 1) {
+			Logger::Error("Error trying to delete %s from previous update (%d)\n", RenamedUpdaterExeFilename, GetLastError());
+		}
+		if (attempts <= 5) {
+			std::this_thread::sleep_for(std::chrono::milliseconds(200));
+			continue;
+		}
+		SetCurrentUpdaterState(UPDATER_PROMPTING_USER);
+		int response = MessageBoxA(NULL, "Failed to delete REPENTOGONLauncherUpdater.exe.bak file from prior update.\n"
+			"The update was most likely successful, but please report this as a bug.\n"
+			"Select \"cancel\" to allow the Launcher to start.\n", UpdaterProcessName, MB_ICONINFORMATION | MB_RETRYCANCEL);
+		if (response == IDRETRY || response == IDOK || response == IDTRYAGAIN) {
+			std::this_thread::sleep_for(std::chrono::milliseconds(200));
+			continue;
+		}
+		Logger::Error("Failed to delete %s from previous update, and the user chose to cancel.\n", RenamedUpdaterExeFilename);
+		return false;
+	}
+
+	return true;
+}
+
+bool Updater::VerifyLauncherNotRunning() {
+	Utils::ScopedHandle fh(INVALID_HANDLE_VALUE);
+
+	int attempts = 0;
+	while (std::filesystem::exists(LauncherExePath) && (fh = Utils::ScopedHandle(CreateFileA(LauncherExePath, GENERIC_READ | GENERIC_WRITE, 0, NULL, OPEN_EXISTING, 0, NULL))) == INVALID_HANDLE_VALUE) {
+		attempts++;
+		if (attempts == 1) {
+			Logger::Error("The launcher exe is currently locked!\n");
+		}
+		if (attempts <= 5) {
+			std::this_thread::sleep_for(std::chrono::milliseconds(200));
+			continue;
+		}
+		int response = MessageBoxA(NULL, "Cannot update the REPENTOGON Launcher, as it is currently running.\nPlease close it and try again.\n", UpdaterProcessName, MB_ICONINFORMATION | MB_RETRYCANCEL);
+		if (response == IDRETRY || response == IDOK || response == IDTRYAGAIN) {
+			std::this_thread::sleep_for(std::chrono::milliseconds(500));
+			continue;
+		}
+		Logger::Error("The launcher is currently running, and the user chose to cancel.\n");
+		return false;
+	}
+
+	return true;
+}
+
 std::unique_ptr<Updater::UniqueWindow> Updater::CreateProgressBarWindow() {
 	HINSTANCE hInstance = GetModuleHandle(NULL);
 
@@ -204,6 +255,9 @@ void Updater::ProgressBarThread() {
 }
 
 void Updater::StartLauncher() {
+	Updater::SetCurrentUpdaterState(Updater::UPDATER_STARTING_LAUNCHER);
+	Logger::Info("Starting launcher...\n");
+
 	STARTUPINFOA startup;
 	memset(&startup, 0, sizeof(startup));
 
@@ -220,58 +274,53 @@ void Updater::StartLauncher() {
 	}
 }
 
-int APIENTRY WinMain(HINSTANCE, HINSTANCE, LPSTR cli, int) {
-	Logger::Init("updater.log");
-
-	Logger::Info("%s started with command-line args: %s\n", Updater::UpdaterProcessName, cli);
-
-	// Parse command-line
+Updater::UpdateLauncherResult Updater::TryUpdateLauncher(LPSTR cli) {
+	// Parse command-line inputs.
 	int argc = 0;
-	char** argv = Updater::Utils::CommandLineToArgvA(cli, &argc);
+	char** argv = Utils::CommandLineToArgvA(cli, &argc);
 
 	// Open a handle for the active launcher process, if provided by the launcher itself.
-	HANDLE launcherHandle = Updater::TryOpenLauncherProcessHandle(argc, argv);
+	HANDLE launcherHandle = TryOpenLauncherProcessHandle(argc, argv);
 
 	// Validate that no other instance of the updater is running.
 	HANDLE lockFile = NULL;
 
-	Updater::Utils::LockUpdaterResult lockResult = Updater::Utils::LockUpdater(&lockFile);
+	Utils::LockUpdaterResult lockResult = Utils::LockUpdater(&lockFile);
 
-	if (lockResult != Updater::Utils::LOCK_UPDATER_SUCCESS) {
-		bool isForced = Updater::Utils::IsForced(argc, argv);
+	if (lockResult != Utils::LOCK_UPDATER_SUCCESS) {
+		bool isForced = Utils::IsForced(argc, argv);
 
 		if (!isForced) {
-			if (lockResult == Updater::Utils::LOCK_UPDATER_ERR_INTERNAL) {
+			if (lockResult == Utils::LOCK_UPDATER_ERR_INTERNAL) {
 				MessageBoxA(NULL, "Unable to check if another instance of the updater is already running\n"
-					"Check the log file (you may need to launch Isaac normaly at least once; otherwise rerun with --force)\n", Updater::UpdaterProcessName, MB_ICONERROR);
-				Logger::Fatal("Error while attempting to lock the updater\n");
-			} else {
+					"Check the log file (you may need to launch Isaac normaly at least once; otherwise rerun with --force)\n", UpdaterProcessName, MB_ICONERROR);
+				Logger::Error("Error while attempting to lock the updater\n");
+			}
+			else {
 				MessageBoxA(NULL, "Another instance of the updater is already running, "
 					"terminate it first then restart the updater\n"
-					"(If no other instance of the updater is running, rerun with --force)\n", Updater::UpdaterProcessName, MB_ICONERROR);
-				Logger::Fatal("Launcher updater is already running, terminate other instances first\n");
+					"(If no other instance of the updater is running, rerun with --force)\n", UpdaterProcessName, MB_ICONERROR);
+				Logger::Error("Launcher updater is already running, terminate other instances first\n");
 			}
 
-			return -1;
-		} else {
+			return UPDATE_ERROR;
+		}
+		else {
 			Logger::Warn("Cannot take updater lock, but ignoring (--force given)\n");
 		}
 	}
 
 	Updater::Utils::ScopedHandle scopedHandle(lockFile);
 
-	// Start up a progress bar window on a separate thread so that we can show some not-frozen UI to the user.
-	std::thread progressBarThread(Updater::ProgressBarThread);
-
-	Updater::SetCurrentUpdaterState(Updater::UPDATER_CHECKING_FOR_UPDATE);
+	SetCurrentUpdaterState(UPDATER_CHECKING_FOR_UPDATE);
 
 	// Check for available updates.
 	// TODO: Properly hook up logging from LauncherUpdateManager
 	NopLogGUI gui;
-	using lu = Updater::LauncherUpdateManager;
+	using lu = LauncherUpdateManager;
 	lu updateManager(&gui);
-	
-	bool allowUnstable = Updater::Utils::AllowUnstable(argc, argv);
+
+	bool allowUnstable = Utils::AllowUnstable(argc, argv);
 	if (allowUnstable) {
 		Logger::Info("Fetching unstable releases is enabled.\n");
 	}
@@ -282,71 +331,54 @@ int APIENTRY WinMain(HINSTANCE, HINSTANCE, LPSTR cli, int) {
 	switch (checkResult) {
 	case lu::SELF_UPDATE_CHECK_UP_TO_DATE:
 		Logger::Info("launcher is already up-to-date.\n");
-		Updater::StartLauncher();
-		return 0;
+		return UPDATE_ALREADY_UP_TO_DATE;
 
 	case lu::SELF_UPDATE_CHECK_ERR_GENERIC:
-		MessageBoxA(NULL, "An error was encountered while checking for the availability of updates. Check the log file for more details.\n", Updater::UpdaterProcessName, MB_ICONERROR);
-		Logger::Fatal("Error while checking for updates.\n");
-		return -1;
+		SetCurrentUpdaterState(UPDATER_FAILED);
+		MessageBoxA(NULL, "An error was encountered while checking for the availability of updates. Check the log file for more details.\n", UpdaterProcessName, MB_ICONERROR);
+		Logger::Error("Error while checking for updates.\n");
+		return UPDATE_ERROR;
 
 	case lu::SELF_UPDATE_CHECK_UPDATE_AVAILABLE:
 		Logger::Info("Found available update version: %s (from %s)\n", version.c_str(), url.c_str());
 		break;
 
 	default:
-		Logger::Fatal("Recieved unexpected error code %d while checking for availability of self updates. Please report this as a bug.\n", checkResult);
-		return -1;
+		SetCurrentUpdaterState(UPDATER_FAILED);
+		MessageBoxA(NULL, "An unknown error was encountered while checking for the availability of updates. Check the log file for more details.\n", UpdaterProcessName, MB_ICONERROR);
+		Logger::Error("Recieved unexpected error code %d while checking for availability of self updates. Please report this as a bug.\n", checkResult);
+		return UPDATE_ERROR;
 	}
 
 	// Check for a renamed updater exe from a previous update and try to delete it if found.
 	// Since we have the lockfile to prevent multiple instances of the updater running at the same time,
 	// we shouldn't ever fail to delete it, but you never know...
-	if (std::filesystem::exists(Updater::RenamedUpdaterExeFilename)) {
-		int attempts = 0;
-		while (std::filesystem::exists(Updater::RenamedUpdaterExeFilename) && !DeleteFileA(Updater::RenamedUpdaterExeFilename)) {
-			attempts++;
-			if (attempts == 1) {
-				Logger::Error("Error trying to delete %s from previous update (%d)\n", Updater::RenamedUpdaterExeFilename, GetLastError());
-			}
-			if (attempts <= 5) {
-				std::this_thread::sleep_for(std::chrono::milliseconds(200));
-				continue;
-			}
-			Updater::SetCurrentUpdaterState(Updater::UPDATER_PROMPTING_USER);
-			int response = MessageBoxA(NULL, "Failed to delete REPENTOGONLauncherUpdater.exe.bak file from prior update.\n"
-				"The update was most likely successful, but please report this as a bug.\n"
-				"Select \"cancel\" to allow the Launcher to start.\n", Updater::UpdaterProcessName, MB_ICONINFORMATION | MB_RETRYCANCEL);
-			if (response == IDRETRY || response == IDOK || response == IDTRYAGAIN) {
-				std::this_thread::sleep_for(std::chrono::milliseconds(200));
-				continue;
-			}
-			Logger::Fatal("Failed to delete %s from previous update, and the user chose to cancel.\n", Updater::RenamedUpdaterExeFilename);
-			return -1;
-		}
+	if (!TryDeleteOldRenamedUpdaterExe()) {
+		return UPDATE_ERROR;
 	}
 
-	Updater::SetCurrentUpdaterState(Updater::UPDATER_PROMPTING_USER);
+	SetCurrentUpdaterState(UPDATER_PROMPTING_USER);
 
 	// An appropriate update is available. Prompt the user about it.
 	const std::string updatePrompt = (std::ostringstream() << "An update is available for the launcher.\n" <<
 		"It will update from version " << Launcher::version << " to version " << version << ".\n\n" <<
 		"Do you want to update now?").str();
 
-	int updatePromptResult = MessageBoxA(NULL, updatePrompt.c_str(), Updater::UpdaterProcessName, MB_ICONINFORMATION | MB_YESNO);
+	int updatePromptResult = MessageBoxA(NULL, updatePrompt.c_str(), UpdaterProcessName, MB_ICONINFORMATION | MB_YESNO);
 	if (updatePromptResult != IDOK && updatePromptResult != IDYES) {
 		Logger::Info("Launcher updated rejected by user\n");
-		return 0;
+		return UPDATE_SKIPPED;
 	}
 
-	Updater::SetCurrentUpdaterState(Updater::UPDATER_DOWNLOADING_UPDATE);
+	SetCurrentUpdaterState(UPDATER_DOWNLOADING_UPDATE);
 
 	// If we opened a launcher process handle, terminate it now.
 	if (launcherHandle != INVALID_HANDLE_VALUE) {
 		Logger::Info("Terminating launcher...\n");
 		if (TerminateProcess(launcherHandle, 0)) {
 			WaitForSingleObject(launcherHandle, INFINITE);
-		} else {
+		}
+		else {
 			// We continue on after this error since we'll notice if the launcher exe is locked and prompt the user about it then.
 			Logger::Error("Failed to terminate launcher process (%d)\n", GetLastError());
 		}
@@ -356,43 +388,18 @@ int APIENTRY WinMain(HINSTANCE, HINSTANCE, LPSTR cli, int) {
 	Logger::Info("Starting update download from %s\n", url.c_str());
 
 	if (!updateManager.DownloadAndExtractUpdate(url.c_str())) {
-		Updater::SetCurrentUpdaterState(Updater::UPDATER_FAILED);
-		MessageBoxA(NULL, "Failed to download the update for the launcher. Check the log file for more details.\n", Updater::UpdaterProcessName, MB_ICONERROR);
-		Logger::Fatal("Error trying to download update from URL: %s\n", url.c_str());
-		return -1;
+		SetCurrentUpdaterState(UPDATER_FAILED);
+		MessageBoxA(NULL, "Failed to download the update for the launcher. Check the log file for more details.\n", UpdaterProcessName, MB_ICONERROR);
+		Logger::Error("Error trying to download update from URL: %s\n", url.c_str());
+		return UPDATE_ERROR;
 	}
 
-	Updater::SetCurrentUpdaterState(Updater::UPDATER_INSTALLING_UPDATE);
+	SetCurrentUpdaterState(UPDATER_INSTALLING_UPDATE);
 
-	// Check if the updater exe is still locked. If so, prompt the user about it.
+	// Check if the launcher exe is still locked. If so, prompt the user about it.
 	// The exe file can remain locked very briefly after terminating the launcher, so we check this after the download.
-	if (std::filesystem::exists(Updater::LauncherExePath)) {
-		bool updaterExeMaybeLocked = true;
-		int attempts = 0;
-		while (updaterExeMaybeLocked) {
-			attempts++;
-			HANDLE fh = CreateFileA(Updater::LauncherExePath, GENERIC_READ | GENERIC_WRITE, 0, NULL, OPEN_EXISTING, 0, NULL);
-			if (fh == INVALID_HANDLE_VALUE) {
-				if (attempts == 1) {
-					Logger::Error("The launcher exe is currently locked!\n");
-				}
-				if (attempts <= 5) {
-					std::this_thread::sleep_for(std::chrono::milliseconds(200));
-					continue;
-				}
-				int response = MessageBoxA(NULL, "Cannot update the REPENTOGON Launcher, as it is currently running.\nPlease close it and try again.\n", Updater::UpdaterProcessName, MB_ICONINFORMATION | MB_RETRYCANCEL);
-				if (response == IDRETRY || response == IDOK || response == IDTRYAGAIN) {
-					std::this_thread::sleep_for(std::chrono::milliseconds(500));
-					continue;
-				}
-				Logger::Fatal("The launcher is currently running, and the user chose to cancel.\n");
-				return -1;
-			} else {
-				updaterExeMaybeLocked = false;
-				CloseHandle(fh);
-				break;
-			}
-		}
+	if (!VerifyLauncherNotRunning()) {
+		return UPDATE_ERROR;
 	}
 
 	// The updater cannot directly overwrite its own currently-running exe with the new one from the update.
@@ -401,30 +408,56 @@ int APIENTRY WinMain(HINSTANCE, HINSTANCE, LPSTR cli, int) {
 	//   2. Make a copy of the existing exe under the original name, just in case we encounter an error during unpacking and can't write the new exe.
 	// So the result of this is that the running exe has been renamed as `exe.bak` (and is still locked by windows),
 	// AND we have a copy under the original name (which is not locked and can be overwritten).
-	if (std::filesystem::exists(Updater::UpdaterExeFilename)) {
-		std::filesystem::rename(Updater::UpdaterExeFilename, Updater::RenamedUpdaterExeFilename);
-		std::filesystem::copy(Updater::RenamedUpdaterExeFilename, Updater::UpdaterExeFilename);
+	if (std::filesystem::exists(UpdaterExeFilename)) {
+		std::filesystem::rename(UpdaterExeFilename, RenamedUpdaterExeFilename);
+		std::filesystem::copy(RenamedUpdaterExeFilename, UpdaterExeFilename);
 	}
 
 	if (!Unpacker::ExtractArchive(Comm::UnpackedArchiveName)) {
-		Updater::SetCurrentUpdaterState(Updater::UPDATER_FAILED);
-		MessageBoxA(NULL, "Unable to unpack the update, check log file\n", Updater::UpdaterProcessName, MB_ICONERROR);
-		Logger::Fatal("Error while extracting archive\n");
-		return -1;
+		SetCurrentUpdaterState(UPDATER_FAILED);
+		MessageBoxA(NULL, "Unable to unpack the update, check log file\n", UpdaterProcessName, MB_ICONERROR);
+		Logger::Error("Error while extracting archive\n");
+		return UPDATE_ERROR;
 	}
 
 	if (!DeleteFileA(Comm::UnpackedArchiveName)) {
-		Updater::SetCurrentUpdaterState(Updater::UPDATER_FAILED);
+		SetCurrentUpdaterState(UPDATER_FAILED);
 		MessageBoxA(NULL, "Unable to delete the archive containing the update.\n"
-			"Delete the archive (launcher_update.bin) then you can start the launcher\n", Updater::UpdaterProcessName, MB_ICONERROR);
-		Logger::Fatal("Error while deleting archive (%d)\n", GetLastError());
+			"Delete the archive (launcher_update.bin) then you can start the launcher\n", UpdaterProcessName, MB_ICONERROR);
+		Logger::Error("Error while deleting archive (%d)\n", GetLastError());
+		return UPDATE_ERROR;
+	}
+
+	Logger::Info("Update successful!\n");
+	return UPDATE_SUCCESSFUL;
+}
+
+int APIENTRY WinMain(HINSTANCE, HINSTANCE, LPSTR cli, int) {
+	Logger::Init("updater.log");
+
+	Logger::Info("%s started with command-line args: %s\n", Updater::UpdaterProcessName, cli);
+
+	// Start up a progress bar window on a separate thread so that we can show some not-frozen UI to the user.
+	std::thread progressBarThread(Updater::ProgressBarThread);
+
+	Updater::UpdateLauncherResult result = Updater::TryUpdateLauncher(cli);
+
+	switch (result) {
+	case Updater::UPDATE_ALREADY_UP_TO_DATE:
+	case Updater::UPDATE_SKIPPED:
+		return 0;
+
+	case Updater::UPDATE_SUCCESSFUL:
+		Updater::StartLauncher();
+		return 0;
+
+	case Updater::UPDATE_ERROR:
+		return -1;
+
+	default:
+		Logger::Error("Unexpected update result value: %d", result);
 		return -1;
 	}
 
-	Updater::SetCurrentUpdaterState(Updater::UPDATER_STARTING_LAUNCHER);
-
-	Logger::Info("Update successful. Starting launcher...\n");
-	Updater::StartLauncher();
-
-	return 0;
+	return -1;
 }
